requirements.txt
pydantic
langchain_community 
tiktoken  
langchainhub 
chromadb 
langchain 
langgraph  
google-generativeai
langchain-google-genai
langchain-groq
langchain-nomic
langchain_mistralai
langchain_huggingface
google-search-results
faiss-cpu
ipykernel
ipywidgets
python-dotenv
youtube_search
streamlit
pillow
pytesseract
speechrecognition
playsound
gtts
# -e .
mysql-connector-python 
flask 

# fastapi 
# uvicorn
fastapi[standard]
uvicorn[standard]
# streamlit 

# uuid # for react 
# event-source-polyfill # for react
# react-icons